chore(DashboardHeader) convert to TS
