+++ packages/react-core/src/demos/examples/DashboardWrapper.js
@@ -13,7 +13,7 @@ import {
   Text,
   TextContent
 } from '@patternfly/react-core';
-import DashboardHeader from '@patternfly/react-core/src/demos/examples/DashboardHeader';
+import DashboardHeader from '@patternfly/react-core/src/demos/examples/Page/DashboardHeader';
 
 export const DashboardBreadcrumb = (
   <Breadcrumb>
